pyproject.toml
[project]
name = "googleadsparser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.9.0",
    "httpx[socks]>=0.28.1",
    "isort>=7.0.0",
    "openai>=2.5.0",
    "pillow>=12.0.0",
    "psutils>=3.3.14",
    "pydantic-settings>=2.11.0",
    "pytesseract>=0.3.13",
    "uiautomator2>=3.4.2",
]
